Refactor: fix type mismatches, remove unused routes, add error pages, clean up anti-patterns
Agent-Logs-Url: https://github.com/sublimit1736/17SimpleBlog/sessions/a3905187-4f0a-4b3d-bfc2-ead66c40c57a

+++ SimBlog17API/src/main/java/cn/chunana/simblog17api/dto/response/ArticleResponse.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @Builder
 @Schema(description = "Article response with full content")
 public record ArticleResponse(
-        @Schema(description = "Article id", example = "1") Integer id,
+        @Schema(description = "Article id", example = "1") Long id,
         @Schema(description = "Article title", example = "My first article") String title,
         @Schema(description = "Article content") String content,
         @Schema(description = "Article content type", example = "MARKDOWN") String contentType,
